backend/requirement.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
python-dotenv
httpx
mangum
PyJWT[crypto]
slowapi
python-multipart
sqlalchemy
python-telegram-bot
resend